alterações da aula
colocado as alterações da aula que estava no bkp para lista.c
